requirements.txt
fastapi
uvicorn
python-dotenv
openai
whisper
ffmpeg-python
pydantic